Add files via upload
Here using the Tutorial, two examples of code have been written for creating a Union. 
The first showing an ideal instance. Data types the same with the same number of columns. 
The second showing a less Ideal instance: 
Although the columns were selected in the statement, there can be some issues with creating the UNION when the tables are merged together. Instances such as the AGE in the file can affect the insights. 

Considerations will have to be made on when and where to use a UNION statement over a JOIN statement